force to compile
the high precision in the mock should be u128
but at the moment setting it to that fails to compile.

+++ substrate/frame/transaction-payment/asset-conversion-tx-payment/src/mock.rs
@@ -251,7 +251,7 @@ impl pallet_asset_conversion::Config for Runtime {
 	type MintMinLiquidity = ConstU64<100>; // 100 is good enough when the main currency has 12 decimals.
 
 	type Balance = u64;
-	type HigherPrecisionBalance = u128;
+	type HigherPrecisionBalance = u64;
 
 	type MultiAssetId = NativeOrAssetId<u32>;
 	type MultiAssetIdConverter = NativeOrAssetIdConverter<u32>;
